Created toll-calculator directory
