--- src/main/java/com/tollcalculator/boObject/TollCalculatorResponse.java
@@ -1,15 +1,0 @@
-package com.tollcalculator.boObject;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.math.BigDecimal;
-
-@Getter
-@Setter
-@Builder
-public class TollCalculatorResponse {
-
-    private BigDecimal taxAmount;
-}
