--- src/main/java/com/tollcalculator/model/Vehicle.java
@@ -1,25 +1,0 @@
-package com.tollcalculator.model;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-@Setter
-@Getter
-@Table(name = "vehicle")
-public class Vehicle {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    @Column(name = "name")
-    private String name;
-
-    @ManyToMany(mappedBy = "taxExemptVehicles")
-    private Set<City> citySet = new HashSet<>();
-}
