+++ toll-calculator/src/main/java/com/tollcalculator/boObject/VehicleType.java
