+++ toll-calculator/src/main/resources/application.properties
@@ -1,0 +1,25 @@
+spring.liquibase.change-log=classpath:/dblogs/changelog/changelog-master.xml
+logging.level.liquibase = INFO
+
+#MYSQL database local
+spring.datasource.url=jdbc:mysql://localhost:3306/toll_fee?createDatabaseIfNotExist=true
+spring.datasource.username=mysql
+spring.datasource.password=passpass
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
+
+#MYSQL database free cloud
+#spring.datasource.url=jdbc:mysql://sql8.freemysqlhosting.net/sql8618423?createDatabaseIfNotExist=true
+#spring.datasource.username=sql8618423
+#spring.datasource.password=w4pHA3S3L6
+#spring.jpa.hibernate.ddl-auto=validate
+#spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
+
+#test if using context feature of Liquibase
+spring.liquibase.contexts = toll_calculator
+spring.liquibase.parameters.key = LiquibaseWithSpringBoot
+
+springdoc.swagger-ui.path=/swagger-ui.html
+springdoc.api-docs.path=/api-docs
+
+logging.level.root=INFO
